iio: magnetometer: yas530: Add YAS537 variant
This adds support for the magnetometer Yamaha YAS537. The additions are based
on comparison of Yamaha Android kernel drivers for YAS532 [1] and YAS537 [2].

Functions used by YAS530 & YAS532 only were renamed from yas5xx to yas530_532.
Registers were renamed accordingly.

In the Yamaha YAS537 Android driver, there is an overflow/underflow control
implemented. It's hard to distinguish if this is for factory testing yas_ext()
function or for extreme cases. For regular usage, this seems not necessary. A
similar overflow/underflow control of Yamaha YAS530/532 Android driver isn't
integrated in the mainline driver. It is therefore skipped for YAS537 in
mainline too.

Also in the Yamaha YAS537 Android driver, at the end of the reset_yas537()
function, a measurement is saved in "last_after_rcoil". Later on, this is
compared to current measurements. If the difference gets too big, a new
reset is intialized. Again it's hard to say if this is necessary for regular
operation, the difference in measurements needs to be quite big. Therefore this
isn't integrated in the mainline driver either.

[1] https://github.com/msm8916-mainline/android_kernel_qcom_msm8916/blob/GT-I9195I/drivers/iio/magnetometer/yas_mag_drv-yas532.c
[2] https://github.com/msm8916-mainline/android_kernel_qcom_msm8916/blob/GT-I9195I/drivers/iio/magnetometer/yas_mag_drv-yas537.c