iio: magnetometer: yas530: Remove redundant defaults on switch devid
This is a preparation for adding YAS537 variant.

In function yas5xx_probe(), there is a switch statement checking for device
IDs. If the ID is unknown, it exits with a device error. In later functions,
it's not neccessary to check the validity of the device IDs again.

When adding YAS537 in a later patch, several of such switch statements will be
added. To make it more uniform, the redundant ones in YAS530/532 get herby
removed. This is done in a separate patch for better history control.