power: supply: rt5033_battery: Adopt status property from charger
The rt5033-battery fuelgauge can't get a status by itself. The rt5033-charger
can, so let's get this value.

The status property of the rt5033-charger, on the other hand, doesn't seem to
reach a status "full" in practice. In the downstream Android driver this state
gets set by end-of-charge IRQ handler [1][2]. In the rt5033-battery driver,
status "full" can be set when reaching a capacity of 100 %.

[1] https://github.com/msm8916-mainline/linux-downstream/blob/GT-I9195I/drivers/battery/rt5033_charger.c#L1082
[2] https://github.com/msm8916-mainline/linux-downstream/blob/GT-I9195I/drivers/battery/rt5033_charger.c#L669-L670

Signed-off-by: Jakob Hauser <jahau@rocketmail.com>